requirements.txt
robotframework==5.0
robotframework-appiumlibrary==1.6.3
pyyaml==6.0